Move datastore proptypes and add list type
This commit creates a new file, next to datastore.js, named
"datastore-proptypes.js". This includes all the previous prop types
offered (node, nodeId, nodeOrId) as well as a new type, named
"listOfNodesOrElements".

"listOfNodesOrElements" recursively scans an array and checks that each
element is either a node, a node id, or some renderable value (strings,
React elements, etc).

Files that previously imported NodePropTypes from datastore.js were
changed prior to this commit.
