+++ src/noteboard/datastore.js
@@ -384,46 +384,3 @@ class NodeList {
         this._readOnlyArr = null
     }
 }
-
-function createRequireablePropType(typeFn, typeName) {
-    const type = (props, propName, componentName) => {
-        const prop = props[propName]
-        if (prop === undefined || prop === null) {
-            return null
-        } else {
-            return typeFn(props, propName, componentName)
-        }
-    }
-    
-    type.isRequired = (props, propName, componentName) => {
-        const prop = props[propName]
-        if (prop === undefined || prop === null) {
-            return new Error(`Invalid prop '${propName}' supplied to ${componentName}; a ${typeName} was required`)
-        } else {
-            return typeFn(props, propName, componentName)
-        }
-    }
-    
-    return type
-}
-
-export const NodePropTypes = {
-    node: createRequireablePropType((props, propName, componentName) => {
-        const node = props[propName]
-        if (!nodeStore.isNode(node)) {
-            return new Error(`Invalid prop '${propName}' supplied to ${componentName}; expected a valid node, got '${node}'`)
-        }
-    }, "node"),
-    nodeId: createRequireablePropType((props, propName, componentName) => {
-        const nodeId = props[propName]
-        if (!nodeStore.isNodeId(nodeId)) {
-            return new Error(`Invalid prop '${propName}' supplied to ${componentName}; expected a valid nodeId, got '${nodeId}'`)
-        }
-    }, "nodeId"),
-    nodeOrId: createRequireablePropType((props, propName, componentName) => {
-        const nodeOrId = props[propName]
-        if (!nodeStore.isNode(nodeOrId) && !nodeStore.isNodeId(nodeOrId)) {
-            return new Error(`Invalid prop '${propName}' supplied to ${componentName}; expected a valid node or nodeId, got '${nodeOrId}'`)
-        }
-    }, "nodeOrId"),
-}
